successfully solved the parameter problem when retrieving scholarship data.
In this commit, I changed the use of parameters in retrieving scholarship data to using a query. This is done because using a query is more flexible than using parameters. In addition, in this commit, routing has also been made to retrieve scholarship data based on order. Then, all data will only be displayed with a LIMIT of 10 which aims to maintain website performance, this LIMIT is set in the model related to scholarships.