Change the routing, controller, and model of study programs, clusters, and universities using sequelize
In this commit, all routing, controllers, and models of study programs, clusters, and universities that previously had each are now merged into one with the name 'profilJurusan' for both routing, controllers, and models. This aims to facilitate future maintenance and explain the existing features. If we continue to use independent routing, controllers, and models for study programs, clusters, and universities, it is feared that it will be difficult to integrate the data.